+++ tests/Tests.Objects/TestJsonContext.cs
@@ -2,5 +2,6 @@ using System.Text.Json.Serialization;
 
 namespace Tests;
 
-[JsonSerializable(typeof(ObjectWithSimpleProperties))]
+[JsonSerializable(typeof(ObjectWithStringList))]
+[JsonSerializable(typeof(ObjectWithTwoScalars))]
 internal partial class TestJsonContext : JsonSerializerContext;